Extract RequirementsTooltip to a separate class
